add pyproject + cli
